chore: updating naming and formatting for clarity
